requirements.txt
python-dotenv
google-generativeai
python-telegram-bot[ext]
schedule
watchdog
dependency-injector
